requirements.txt
# API server
fastapi
uvicorn[standard]
pydantic-settings
python-dotenv
httpx
aiofiles
python-multipart

# RAG pipeline
langchain
langchain-text-splitters
langgraph
qdrant-client
sentence-transformers
rank-bm25
numpy

# Document parsing
unstructured[pdf,docx]
pymupdf
python-docx

# LLM + Embedding (one package for both)
openai

# Evaluation
ragas
datasets

# Cache & DB
redis
